Move ENV setup into spec_helper.rb
Previously, the tests were dependent on the spec/support files being
loaded in a order (specifically, the env setup had to be loaded before
the router helpers).  This order is not guaranteed, so moving the env
setup into the spec helper will ensure it's done before the router
helpers are loaded etc.