Add a document describing the content of documents
This gives details of the formats and semantics of fields included in
documents.

I wrote this partly in an attempt to ensure that I understand the
semantics of this, and partly to get some good documentation around this
for others. There are still some parts of this I'm uncertain about, so
I've marked them with TODO.