Rename the new version of tags to `links`.
This is as a result of a few discussions about potential naming
conventions, and the benefits and disadvantages of each. Most of the
disadvantages stem from clashes with existing terms, and the potential
for confusion.

In the world of Panopticon <https://github.com/alphagov/panopticon> and
the content API <https://github.com/alphagov/govuk_content_api>, tags
are treated as first-class objects, with their own attributes and
persistence; in the content store, we're using them to express a
typed relationship between two pieces of content. This may conflict with
existing mental models, both of tags as they exist in the GOV.UK
publishing system, and of tags in the wider world as keywords
designating a subset of objects.

Some possibilities we've considered:

  * `tags` - clashes with existing naming convention, and the existing
    concept of tags on the Web

  * `links` - clashes with the concept of hyperlinks, although our
    planned implementation maps to HTML links and the `rel` attribute
    <http://www.w3.org/TR/html5/links.html>

  * `related` - clashes with the fact that we're expecting `related` to
    be one of the link types we use

  * `associations` - on the plus side, no pre-existing mental model to
    clash with; on the minus side, no pre-existing mental model at all