Add optional content IDs to the store.
Once all the client apps are generating content IDs, we can make these
mandatory.

The UUID validation code is taken directly from the contacts app
<https://github.com/alphagov/hmrc-contacts/commit/98a099dee004>. It's
not worth the extra time right now to pull this out into a shared gem,
but it's something worth doing later if/when we want to use this in a
third place, or if we need to update the validation logic.