Refer to this app consistently.
We haven't had a consistent naming convention so far, so several
versions have cropped up:

  * Content Store
  * the content store
  * content-store
  * the content-store

I've rejected the two hyphenated versions in the name of plain English.
The choice of "the content store" over "Content Store" is
semi-arbitrary: I think it reflects how we tend to refer to this app,
but that may well reflect confirmation bias on my part.