Change the documentation around link types.
The previous documentation said that the keys represented the type of
the linked item (`organisations` or `topics`, for example); it's more
accurate to say that it's the type of the relationship itself.

For example, the `related` link type doesn't care about the type of the
thing it's linking to. As anotehr example, we could feasibly add `next`
or `previous` links for a series of items (multi-part guides, say), or
`up` links to a parent section.