Be explicit about single-item relationships.
For example, a `primary_section` relationship might only make sense with
a single linked item, but for the sake of consistency of representation,
we're sticking with a list of 1 item.

If someone were to implement this incorrectly, this would get caught and
rejected by the model logic anyway, but since people have asked about
this behaviour before, it's worth adding it to the documentation so we
can cite it.