Enable message queue by default in development.
The current setting whereby it's disabled by default has led to
confusion.  Given rabbitmq is available and configured on the dev VM it
makes sense for this to be enabled by default.

It remains disabled by default for the tests because the test quite
enables it explicitly for the tests that need it.