Adds HeartbeatGenerator and GovukExchange
For monitoring purposes, a heartbeat message is sent to the
published_documents exchange. These messages have a timestamp and the
hostname of the machine, along with a header of
"application/x-heartbeat".

This work is performed by the `HeartbeatGenerator` class.

This PR also adds `GovukExchange`, which tries to encapsulate how we
write to exchanges at the moment. It also allows for a shorter setup for
publishing messages, useful for Rake tasks or cron jobs.