Move note about picking latest item for duplicates
This note is only relevant when expanding links, so the note should
belong in that section rather than the section explaining what the
content ID is.  It's misleading otherwise, because when there are
duplicates `content-id`-`locale` pairs, we will still serve the older
one if its path is requested: it's only in link expansion that we
de-dupe.