Deprecate base_path in body in the publishing context
The documentation implies that the base_path must be included in the request
body for write requests as well as in the URL. This is not true and in fact if
supplied in the body it is discarded, otherwise it is ignored - the base_path
in the URL is used instead.

Leading developers to believe that it is required but not performing any
validation on it seems likely to be a source of errors, so let's get rid of it.

Once all clients have been updated to omit it from the request body, we can
stop treating it as a special case and start rejecting it as we would any other
unexpected field.