Add separate GET route for the public API to use
I'm shortly planning to make the content-store "GET content" API public.  As
part of that, we'd like to be very sure that only GET requests will be routed
from the public API.

To help with this, this commit adds a route to handle GET requests to
`/api/content/foo`.  This route serves exactly the same content as
`/content/foo`, but only handles GET requests.

By pointing publicapi at this route, we get some defence-in-depth that only the
read-only part of the API is available.  We also make it easier to distinguish
requests which have come via the public API from those which come from other
internal apps in log files and in systems such as Kibana.