Don't crash on related links containing hashes
I tried to send a frontend example to the content store and the content
store crashed because the validation algorithm did not check whether the
content item is a string before comparing it to a regexp.

I've added a unit test for this scenario and refactored the validation
code to handle this case.

The change made is to use `#=~` rather than `Regexp#match`. Note that
`#=~` is defined on `Object` to always return false.