Make ContentItem access_limiting aware
It will store an access_limiting hash as is. It also has a method to determine
if it is visible by a given user:

content_item.viewable_by?(user_id)

This method will always return true for non-access-limited items. Those with
access-limiting will only return true if the provided user id is present in the
"users" array.