Fix field overwrite warnings
Currently the following is being logged regularly in production:

```
Overwriting existing field _id in class ContentItem.
```

This is a warning generated by Mongoid when you define a field that's
already been defined as that's often in error.  In our case, this is
intentional because we're redefining the _id field. Adding the
`:overwrite => true` tells Mongoid that this is intentional and silences
the warning.

The `overwrite` param isn't mentioned in the documentation.  I found
mention of it in an issue on github -
https://github.com/mongoid/mongoid/issues/3272