Fix heartbeat integration test.
The previous commit made the rake task use the queue defiend in the
initializer. This was set to noop mode for Rails.env.test?.

This updates the initializer to allow the test mode default to be
overridden with an ENV variable, and uses that ENV var in the
integration test.