Add analytics identifier to root object
Analytics identifiers are short codes which we can use to represent
objects in character restricted environments (like Google Analytics
custom dimensions). Having them on the root lets us easily pull them out
to be part of the links hash so that content can be correctly marked
with all related dimensions.