Remove circularity from fetching_content_item_spec
As part of the refactoring in #130, this integration test was updated to
use the Presenter to generate its expectation. This meant that it
wasn't testing the full stack, and was only testing that the returned
JSON matched what the presenter generated. There was nowhere in the test
suite that actually asserted against the specifics of what the
content-store should return in its JSON response.

This updates it to make assertions about the specifics of the JSON
returned.