Don't include self translation with no content_id
When expanding available_translations for an item without a content_id,
the current behaviour is for an item to return itself. This is
problematic because we want to update the schema to require a content_id
in expanded links.

This therefore changes this behaviour so that no available_translations
are returned for an item without a content_id