Add a rake task that reports on duplicate content IDs
Why?:

We want to eventually use content IDs to uniquely identify content in
the publishing API. This report is a starting point for assessing the
different cases where content IDs are non-unique.

How?:

The rake task does a simple count of duplicate content ID occurrences
using a mongo aggregation. It drops a list of content items into a CSV
(displaying a 'useful' subset of a content item's fields) and writes this
to $PROJECT_ROOT/tmp.