Further improve the reporting around duplicate content ids
Why?

This is part of the move to using content IDs as a unique identifier for
every piece of content. Building on top of the existing CSV report,
it'd be useful to understand what the breakdown is of duplicate content
items within the EN locale. This will begin to provide some ideas about
how to approach the migration.

How?

- Extract the report logic into its own class for easier refactoring.
- Provide two different methods to produce a full list of duplicates or
  a list scoped to a given locale.