Add rake task to find inconsistent redirects
In New World Publishing, applications send content to the
publishing-api, which forwards it to the content-store and also
registers routes for the document with the router-api.

However, in the past apps have sent routes directly to the router-api.
This has happened especially when setting up redirects.

It means we have a bunch of routes in the router for which either no
content items exist, or for which the content item doesn't match the
content item.

The rake task added in this PR can be used to find a subset of the last
case. It finds policies with format "placeholder" in the content-store
that are actually redirects. The outputted information will be used to
bring the database up-to-date by turning the content items into
redirect content items. This will be done in a later PR.