Add official redirects for RAIB pages
The pages here were redirected in router-data in
https://github.gds/gds/router-data/commit/15be25b94ef6e0b12f16669d632617
d8005f23a2 because it's "Rail Accident", not "Accidents".

When the redirection happened the corresponding content items have not
been updated. This means that the pages on the new URLs have the same
content_id. This interferes with our work to create a new tagging
infrastructure.

We will remove the redirects from router-data as well.