Remove content items that 404
The paths in the CSV return 404 on the site. For example:

https://www.gov.uk/government/collections/ots-tax-complexity

However, they do have content items.

https://www.gov.uk/api/content/government/collections/ots-tax-complexity

The content_ids for these items are duplicated. In the above case it
duplicates:

https://www.gov.uk/api/content/government/collections/hmt-ots-tax-comple
xity

Because they return 404, we think its best to remove these content
items, as the items have not previously worked.