De-duplicate content items...
Duplicate content items sharing the same content_id and locale
periodically get added to the content store. A detailed explanation
of how this happens can be found here:
https://gov-uk.atlassian.net/wiki/display/FS/Duplicate+content-ids

This pull request adds tasks to find duplicates and report or
de-duplicate by destroying the older record(s) in a duplicate set.