Handle out of order updates...
Since updates to content store are now asynchronous, it is possible for
them to arrive in the wrong order. Content store needs to only persist
the latest update.
The version passed in the payload is used in comparison with the version
on a previous content item, if the existing version is greater than or
equal to the version in the payload the item isn't updated and a 409
conflict is returned.