Backwards compatibility for /v1 payloads…
The /v1 payloads of the publishing api don’t 
enforce that a content_id is present. When this
happens, no version is saved against the content
item in publishing api and this means that a
version is not created for the item. Therefore,
we need to accept payloads to the content store
that don’t have a version.