Use ‘transmitted_at’ rather than ‘version’ for message sequencing…
We initially were planning to use version for
checking if messages have been received in the
correct order. However, we realised that both
content items and link sets have their own
versions. This means that we need to check message
sequence against something unambiguous, i.e. the
transmitted_at timestamp.

This is accurate to within nanoseconds, so it is
extremely unlikely that two messages would have
the same timestamp.

For example:

100.times.map { Time.new.to_f } returns unique
results. Therefore, it is very unlikely these
would ever conflict for actual requests.