Construct linked ContentItems for pass-through links.
This allows the publishing API to provide
"hard-coded" content for linked items in the cases
where we cannot currently represent them as dynamic
items in the publishing API, for example governments
(for history mode).

This is a temporary measure until a publishing API
refactor takes over the expansion of these links
hashes.