Add special case for Topical Events
The start and end dates for Topical Events are needed to render Topical
Event About Pages. We expand these fields in the links hash to simplify
the logic in the frontend.

We need to remove the returned field filter for linked items in order
to achieve this, as changing these `only` lines for each format is not
sustainable.