Remove UpdateLock and transmitted_at
Due to time differences across the host machines of +/- 3ms this
approach is error prone once deployed to integration/staging/production.
This commit just removes the `transmitted_at` field and `UpdateLock`. It
doesn't currently `unset` `transmitted_at` in Mongo. I've not added an
alternative locking approach here as that may well be better added in
the Publishing API.