Only return renderable content for debug endpoint
Currently the debug taggings-per-app endpoint returns the taggings of
all content items, including the (mostly empty) taggings for redirects.
This causes the tagging migration verifier to think that there are
differences between contentapi and content-store.

We are only interested in the taggings of renderable items, not
redirects.

Trello: https://trello.com/c/1QMuBOSN