Basic implementation to pull in policies for working_groups
To render a working_group, you need the list of policies which link to that
working_group.

We could fulfill that by adding special knowledge to the frontend so that it
knows to make a second request, to /incoming-links. There are a few issues with
that:
* one of the design goals of the content-store is that it includes everything
needed to render that page
* it complicates the frontend
* we think that when dependency resolution is implemented, the API of
content-store will change to match this implementation, and that the
/incoming-links endpoint could be removed

If the API design does change in this way, then we don't need to do anything to
our frontends when we implement dependency resolution.

I'd rather it wasn't so directly hardcoded - instead having a list with an
entry for each format, containing the various strings - but I had difficulty
with naming. Probably best to refactor when we get to the second or third use
case.

We are confident that there will be other formats where this functionality will
be required.

I see this change as being similar to 04368e52d22d1e5422608ce18cf11e306014f377.