Fix relative links
A number of links in this document were referring to docs in the `doc`
subdirectory (which is now the file's parent directory).  I suspect this
content may once have been part of the `README` and has since been moved.

The file is at https://github.com/alphagov/content-store/blob/master/doc/technical-information.md