Allow `gone` items to be returned with 200
Currently sending a `gone` item through the `/unpublish` endpoint of
Publishing API always creates a `gone` route in the router and causes
Content Store to also return a 410 status.

Due to the vagueries of Whitehall 'unpublishing' we need a way to
unpublish but still render an explanation and optional alternative url.
This can't be done with a 410 route in the router.

We need to address this properly but for now this change should allow
the item to be rendered correctly and fix the issues we have in
production.