Consider `gone` items with no explanation/alternative_url as 'gone'
A previous PR added a check for the absence of the `details` field on a
`gone` content item and used that to determine whether to consider the
item actually 'gone' (return 410) or gone with content (and return 200).

This commit adds further logic to that to check that there is a `details` field
and that there are values within it. If there are no values we return
410 even if there are keys but they are empty.