Fix a test in the fetching linked items spec
This changes the expectation, from using the fixed current time, to
comparing against the time from the model Additionally, the model is
reloaded to ensure that the representation is the same as the database.

Initially it appeared that my local timezone was leaking in to the test,
and causing the public_updated_at values to differ by an hour in the
actual and expected response. I tried quite a few different ways of
resolving this, and some worked, but led to the times sometimes being a
millisecond apart.

Kevin spotted that calling reload on the ContentItems after they are
created solved the issue with the test and expected value sometimes
being a millisecond apart.