Fix jenkins script
The schema tests are currently not reporting back the
in-progress/success/failure status for a test run, because it expects
jenkins.sh to do this. This repo is using an older version of
jenkins.sh and a separate script for branches (jenkins_branches.sh).

This commit uses the more modern publishing-api script as an example.
It should fix the running of the schema tests.