Allow fact check access via header
If a valid JWT token is passed to authenticating-proxy it will extract
the value to the Govuk-Fact-Check-Id header, which is passed downstream
by gds-api-adapters.

This change then checks that the header value matches a value in the
fact_check_id array within the access_limited hash of the item being
requested and returns it if so, otherwise sends a 403.

Note that the presence of the header means the user has bypassed
normal authentication, so we need to verify that they've specifically
requested the item with that fact_check_id, and not fall through to
the auth check.

As with a normal access-limited item, we send a cache-control: private
header when the request contains the content-id header.