Ensure requests with invalid user id are rejected.
Since authenticating-proxy allows unauthenticated access if any valid
fact_check token ifs passed, it could be possible for a misconfiguration
to fail to pass on the fact_check_id and therefore provide access to
all non-access-limited items.

To prevent this, authenticating-proxy will explicitly set the user
header to 'invalid' if authentication is bypassed. We therefore add a
check here for that value in the case where the fact_check_id does not
match.