Return content when route inside ContentItem matches
This updates the show action to return a ContentItem when the base_path
requested matches a route of a ContentItem rather than just the
base_path.

The reason for this change is that developers working on frontend
applications are having a problem determining which ContentItem they
require for a given path. Eg for /a/b/c they have no means to know that
the content they require is available on /a/b for instance.

This has been added in as part of a quick spike to see if this is
possible and would likely require additional thought on the implications
of using this in content-store.