Always register routes, even if the route set is unchanged
Conditionally registering the routes could lead to routes not being
registered entirely for some content. Consider the following scenario:

 - Start with an empty content store
 - update request made
 - No previous item
 - Item saved to the database
 - Route registration fails
 - Content Store returns error status to client

 - update request retried by client
 - previous_item exists
 - route registration skips, as route_set matches that of the previous
   item
 - Content Store returns success status to client

Even though the 2nd request is "successful", the routes have not been
registered. This can also happen if the item exists in the content
store, but the update has different routes.

Always registering the route seems preferable, as the code is simpler
compared to manually rolling back the creation or update of the local
database, however implementing rollback manually in the code would be
a possible alternative, and this change will result in some additional
redundant requests to the router-api.