BasePathForPath class
This class takes a path and determines the base_path of a content item
for it. If one is not found it returns nil.

It is intended to be used to allow us to look up content items by the
path of a content item that is within the routes of the content item.