Return 303 redirects when base_path doesn't match path
This changes the way ContentItem and PublishIntents are looked up so
that they can match a path (which could be the base_path, or one of the
routes inside the model).

If the base_path of the ContentItem does not match the path requested
then a 303 See Other redirect is returned.