Switch from CGI.escape to ERB url_encode
The switch from URI.escape to CGI.escape had the unfortunate side effect
that the URLs broke when they had spaces in them as CGI.escape uses the
"+" encoding of query strings rather than the %20 used in paths. To get
this we have to use a different method and the ERB::Util.url_encode
seems to be the best option in std lib.