Add register_backends rake task
This provides a convinient way to set backends in the Router API, for
all the content in the Content Store.

This can be useful for a few reasons:
 - You have changed the configuration of the
content store with respect to the backends it generates, and want to
update the Router API.
 - The data in the Router API doesn't match, as you have loaded the
database from another a deployment of these applications with
different configuration.