Log the delay in scheduled publishing
We need data on the time it takes for scheduled publications to reach
the frontend to ensure that we are publishing sufficiently quickly to
meet statuatory requirements.

This commit records the difference between intended publish time and the
actual time that content arrives at content store to statsd.

In order to only log the latency for content submissions that relate to
a PublishIntent, we also delete the PublishIntent if we match one. This
prevents unscheduled submissions later in the day from also recording a
latency. This would give us incorrect data as the latency could be as
long as hours.