Add ScheduledPublishingLogEntry
This object will be persisted when a scheduled publication is saved to
the content store and records base_path, document_type, expected
publication time and actual publication time. It also saves the delay in
milliseconds for convenience.

This will replace the existing statsd logging which is ineffective due
to the low volume of data.

This will also allow us to identify individual items that were delayed
should we need to retrieve this data upon request by publishers in the
event of an issue with publication.