Add rake task to validate a sample of content
The content-store currently returns a lot of invalid items. Before we
start fixing that we need some way to track the number of invalid
things.

This commit adds a rake task that is meant to be run periodically. It
will take 100 items for each schema and validate those. The result is
outputted so that we can see the errors in Jenkins. The overall result
is sent to statsd as a gauge. This means we can generate a graph of the
(sampled) percentage of content items that are invalid.

Ideally we would check every item in the content-store, but this is
prohibitively expensive (it takes about 4 hours). The sample will
hopefully give us enough data to start working on improving the state
of the data.

https://trello.com/c/FesP8XSc